Move __tests__/schema/uuid/thread to __tests__/schema/
